Move BatchClient into the lib folder for future use.
