+++ test/pool-of-servers.js
@@ -22,6 +22,7 @@
 
 var parallel = require('run-parallel');
 
+var BatchClient = require('./lib/batch-client.js');
 var allocCluster = require('./lib/alloc-cluster.js');
 
 allocCluster.test('sending requests to servers synchronously has perfect distribution', {
@@ -85,85 +86,6 @@ allocCluster.test('sending requests to servers synchronously has perfect distrib
         assert.end();
     }
 });
-
-function BatchClient(channel, hosts) {
-    if (!(this instanceof BatchClient)) {
-        return new BatchClient(channel, hosts);
-    }
-
-    var self = this;
-
-    self.channel = channel;
-    self.hosts = hosts;
-
-    self.subChannel = self.channel.makeSubChannel({
-        serviceName: 'server',
-        peers: self.hosts
-    });
-}
-
-BatchClient.prototype.warmUp = function warmUp(callback) {
-    var self = this;
-
-    var waiting = [];
-    for (var l = 0; l < self.hosts.length; l++) {
-        var host = self.hosts[l];
-
-        waiting.push(self.subChannel.waitForIdentified.bind(self.subChannel, {
-            host: host
-        }));
-    }
-
-    parallel(waiting, callback);
-};
-
-BatchClient.prototype.sendRequests = function sendRequests(options, callback) {
-    var self = this;
-
-    var totalRequests = options.totalRequests;
-    var batchSize = options.batchSize;
-
-    var callReqThunks = [];
-    for (var i = 0; i < totalRequests; i++) {
-        var req = self.subChannel.request({
-            serviceName: 'server',
-            hasNoParent: true,
-            timeout: 500,
-            headers: {
-                cn: 'client',
-                as: 'raw'
-            }
-        });
-
-        callReqThunks.push(req.send.bind(req, 'foo', 'a', 'b'));
-    }
-
-    var errorList = [];
-    var resultList = [];
-    (function loop() {
-        if (callReqThunks.length === 0) {
-            return callback(null, {
-                errors: errorList,
-                results: resultList
-            });
-        }
-
-        var parts = callReqThunks.slice(0, batchSize);
-        callReqThunks = callReqThunks.slice(batchSize);
-
-        parallel(parts, onPartial);
-
-        function onPartial(err2, results) {
-            if (err2) {
-                errorList.push(err2);
-            }
-            // assert.ifError(err2, 'expect no req err');
-
-            resultList = resultList.concat(results);
-            loop();
-        }
-    }());
-};
 
 allocCluster.test('sending requests to servers over time has good distribution', {
     numPeers: 26
